Use binutils-2.26 in toolchain & always use --build when building.
Recent config.guess version outputs 'x86_64-pc-linux-gnu' and not
'x86_64-unknown-linux-gnu' for x86_64 Linux systems. This brings in
a possibility that it builds a native tool (in this case, binutils)
instead of a cross one, breaking other build processes. For now, we
will always specify '--build' in configure to fix this, even though
it seemed redundant sometimes.