inittab: Give controlling tty to all ttyN sessions.
The leading hyphen is a BusyBox init feature (read examples/inittab in
BusyBox source code), and it's used in BusyBox's hard-coded default
inittab. There's no reason for us not to use it.

Note: the leading hyphen should NOT be used on /sbin/getty. It's meant
for systems without getty like ours are.